feat: Refactor database entities and update build configurations
This commit introduces significant changes to the database entity relationships and standardizes Maven build configurations across modules.

Database changes:
- `PeerEntity` now has a `@ManyToOne` relationship with `UserEntity`, linking peers directly to users.
- `IpEntity` now has a `@ManyToOne` relationship with `PeerEntity`, ensuring each IP is associated with a peer.
- Updated `@OneToOne` mapping for `IpEntity` in `PeerEntity` to use `id_address_id` as the join column and enabled `orphanRemoval`.
- Changed `UserEntity`'s `id` generation strategy from `AUTO` to `IDENTITY` for better database compatibility and performance.
- Updated `UserEntity`'s `@OneToMany` relationship with `PeerEntity` to use `mappedBy="user"`.

Build configuration updates:
- Added `maven-compiler-plugin` configuration to `agent`, `bot`, `core`, and `db` modules to enable `-parameters` for reflection and configure annotation processors for Lombok and Spring Boot.
- Excluded Lombok from the `spring-boot-maven-plugin` in the `agent` module.

Other changes:
- Renamed the `agent` application to `wireguardbot-agent` in `application.yaml`.
- Updated the `WEBHOOK_BASE_URL` in `addWebhookWireguardBot.sh` to reflect a new path.

+++ common/src/main/java/dev/yonel/wireguardbot/common/services/database/UserService.java
@@ -18,5 +18,5 @@ public interface UserService {
 
     void deleteUser(UserDto user);
 
-    boolean userExistsByUserId(Integer userId);
+    boolean userExistsByUserId(Long userId);
 }
